cleaned up and moved stuff around
